requirements.txt
pycryptodome
pyrebase
beautifulsoup4
python-dotenv
mypy
pycodestyle
pylint